+++ module/macros.js
@@ -396,7 +396,6 @@ export class TorgeternityMacros {
       woundModifier: -_actor.system.wounds.value,
       stymiedModifier: _actor.statusModifiers.stymied,
       vulnerableModifier: _actor.statusModifiers.vulnerable,
-      darknessModifier: _actor.statusModifiers.darkness,
       waitingModifier: _actor.statusModifiers.waiting,
       type: 'skill',
       isOther1: game.scenes.active && game.scenes.active.torg.cosm !== 'none',